Move toHtml to AstVisitorUtils, open visibility for HtmlWriter
Use BigDecimal instead of Decimalformat in AstLinePercentVisitor
